Add federation setup to Prometheus formula (#47)
The change enables configuration of federated Prometheus setup. It assumes matching metric names according to to the specified list of matching patterns. The default is set to:

{__name__=~"node:.*"}

which matches all recording rules with the prefix node:. The corresponding recording rules have to be defined on the cluster's (e.g. CaaSP) Prometheus server.